requirements.txt
# --- Core ---
Flask==3.0.3
gunicorn==21.2.0

# --- Base de données ---
psycopg2-binary==2.9.11
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1

# --- Sécurité et session ---
itsdangerous==2.2.0
Werkzeug==3.0.4

# --- JSON, dates et outils utiles ---
python-dateutil==2.9.0.post0
uuid==1.30
requests==2.32.3

# --- Pour le déploiement Render ---
certifi==2024.8.30
jinja2==3.1.4
